split image and text out into sub components to make stuff easier to read.
