+++ app/assets/javascripts/components.js
@@ -82,6 +82,8 @@ SectionsModalList = require("./components/section/SectionsModalList");
 ItemsList = require("./components/item/ItemsList");
 ItemShow = require("./components/item/ItemShow");
 ItemListItem = require("./components/item/ItemListItem");
+ItemImage = require("./components/item/ItemImage");
+ItemText = require("./components/item/ItemText");
 AdditionalResources = require("./components/item/AdditionalResources");
 AdditionalResourcesItem = require("./components/item/AdditionalResourcesItem");
 ItemsModalList = require("./components/item/ItemsModalList");
